[REFACTORING] ErrorCode를 CommonLib으로 옮기고 관련 내용 정리. Namespace 정리.
