+++ DBServer/DBServer/MongoDBManager.cs
@@ -1,10 +1,9 @@
 ﻿using MongoDB.Driver;
 using System;
-using System.Collections.Generic;
-using System.Text;
 using System.Threading.Tasks;
+using CommonLibrary;
 
-namespace NextManComing_DBServer
+namespace DBServer
 {
 	internal class DBUser
 	{
