+++ Library/LogicLibrary/LogicFunctions.cs
@@ -1,10 +1,7 @@
 ﻿using NetworkLibrary;
 using MessagePack;
-using System;
-using System.Collections.Generic;
-using System.Text;
 using System.Threading.Tasks;
-using NLog;
+using CommonLibrary;
 
 namespace LogicLibrary
 {
