--- Library/ManageLogicLibrary/Class1.cs
@@ -1,8 +1,0 @@
-﻿using System;
-
-namespace ManageLogicLibrary
-{
-    public class Class1
-    {
-    }
-}
