+++ LoginServer/LoginServer/Encrypter.cs
@@ -1,7 +1,5 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Security.Cryptography;
-using System.Text;
 
 namespace LoginServer
 {
