+++ LoginServer/LoginServer/LoginServerConfig.cs
@@ -1,11 +1,8 @@
 ﻿using Newtonsoft.Json.Linq;
 using System;
-using System.Collections.Generic;
 using System.IO;
-using System.Linq;
-using System.Threading.Tasks;
 
-namespace LoginServer 
+namespace LoginServer
 {
 	internal class LoginServerConfig
 	{
