+++ LoginServer/LoginServer/Program.cs
@@ -1,9 +1,6 @@
 ﻿using Microsoft.Owin.Hosting;
 using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Text;
-using System.Threading.Tasks;
 
 namespace LoginServer
 {
