+++ LoginServer/LoginServer/Startup.cs
@@ -1,7 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-using Owin;
+﻿using Owin;
 using System.Web.Http;
 
 namespace LoginServer
